requirements.txt
# ==============================================================================
# File: requirements.txt - Optimized compatible version
# ==============================================================================
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-dotenv==1.0.0
requests==2.31.0

# Core ML/AI packages - compatible versions
numpy<2.0.0
torch
sentence-transformers

# LangChain packages
langchain==0.1.20
langchain-core==0.1.52
langchain-community==0.0.38
langchain-groq==0.1.5
langchain-pinecone==0.1.0
langchain-huggingface

# Document processing
pypdf==3.17.4

# Vector database
pinecone-client==3.2.2